only auth users can post/delete workouts

+++ frontend/src/components/WorkoutDetails.js
@@ -1,3 +1,4 @@
+import { useAuthContext } from "../hooks/useAuthContext"
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
 
 // date fns
@@ -6,10 +7,17 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 const WorkoutDetails = ({ workout }) => { 
 
   const {dispatch} = useWorkoutsContext()
+  const {user} = useAuthContext()
 
   const handleClick = async () => {
+    if(!user){
+      return 
+    }
     const response = await fetch('api/workouts/' + workout._id, {
       method: "DELETE",
+      headers: {
+        'Authorization' : `Bearer ${user.token}`
+      }
     })
 
     const json = await response.json()
